+++ MatrixSDK/MXRestClient.h
@@ -2090,14 +2090,17 @@ FOUNDATION_EXPORT NSString *const kMXMembersOfRoomParametersNotMembership;
                                    failure:(void (^)(NSError *error))failure;
 
 /**
- Get information about the current version.
+ Get information about a backup version.
+
+ @param version the backup version. Nil returns the current backup version.
 
  @param success A block object called when the operation succeeds.
  @param failure A block object called when the operation fails.
 
  @return a MXHTTPOperation instance.
  */
-- (MXHTTPOperation*)keyBackupVersion:(void (^)(MXKeyBackupVersion *keyBackupVersion))success
+- (MXHTTPOperation*)keyBackupVersion:(NSString*)version
+                             success:(void (^)(MXKeyBackupVersion *keyBackupVersion))success
                              failure:(void (^)(NSError *error))failure;
 
 /**